feat: refactor admin panel and user management pages, update navigation links, and remove unused scripts
