+++ js/admin.js
@@ -1,0 +1,5 @@
+import { getUser } from "./users.js"; 
+
+window.addEventListener("DOMContentLoaded", async () => {
+    getUser();
+});
