+++ js/script.js
@@ -1,7 +1,10 @@
-import { TimKiem } from "./SearchService.js";
-import { SaveLocation, GetLoCations } from "./locationService.js"; 
+import { TimKiem } from "./search.js";
+import { SaveLocation, GetLoCations } from "./location.js"; 
 import { auth } from "./config.js";
 import { onAuthStateChanged } from "https://www.gstatic.com/firebasejs/11.3.1/firebase-auth.js";    
+
+
+
 
 const btnSearch = document.getElementById("btn-Search")
 const btnSave = document.getElementById("btn-SaveLocation");
