+++ logs.html
@@ -1,9 +1,9 @@
-<!-- /admin/index.html -->
+<!-- /admin/logs.html -->
 <!DOCTYPE html>
 <html lang="vi">
 <head>
   <meta charset="UTF-8" />
-  <title>Admin - Tổng quan</title>
+  <title>Admin - Lịch sử</title>
   <meta name="viewport" content="width=device-width, initial-scale=1" />
   <link href="https://fonts.googleapis.com/css2?family=Poppins:wght@400;600&display=swap" rel="stylesheet" />
   <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0/dist/css/bootstrap.min.css" rel="stylesheet" />
@@ -14,22 +14,29 @@
     <div class="row">
       <nav class="col-md-3 col-lg-2 sidebar shadow-sm">
         <h4>🌤️ Admin Panel</h4>
-        <a href="admin.html" class="nav-link active">🏠 Tổng quan</a>
-        <a href="logs.html" class="nav-link">🕒 Lịch sử tra cứu</a>
+        <a href="admin.html" class="nav-link">🏠 Tổng quan</a>
+        <a href="logs.html" class="nav-link active">🕒 Lịch sử tra cứu</a>
         <a href="users.html" class="nav-link">👥 Người dùng</a>
         <hr />
         <a id="logout" class="btn btn-sm btn-outline-dark mt-2" href="#">Đăng xuất</a>
       </nav>
 
       <main class="col-md-9 col-lg-10 main">
-        <h2 class="mb-3">📍 Top địa điểm tra cứu</h2>
-        <ul class="top-list" id="top-cities">
-          <!-- Dữ liệu JS đưa vào -->
-        </ul>
+        <h2>🕒 Lịch sử tra cứu</h2>
+        <div class="table-responsive mt-3">
+          <table class="table table-bordered">
+            <thead class="table-light">
+              <tr><th>Người dùng</th><th>Thành phố</th><th>Thời gian</th></tr>
+            </thead>
+            <tbody id="log-table">
+              <!-- JS thêm vào -->
+            </tbody>
+          </table>
+        </div>
       </main>
     </div>
   </div>
   <script type="module" src="js/admin.js"></script>
-  <script type="module" src="js/main.js"></script>
+   <script type="module" src="js/main.js"></script>
 </body>
 </html>
